Tidy up visualization tests
